refactor(cli): migrate registry to TypeScript (#1553)
## Summary

Migrate the sandbox registry implementation from `bin/lib/registry.js`
to `src/lib/registry.ts`, keeping the existing API surface and replacing
the JS file with a shim to the compiled implementation.

## Why

This is the next core-substrate step after #1552:
- move central CLI state management into typed code
- keep the strangler pattern already used across the repo
- unblock later credentials / policies cleanup and TS backedge removal

## Changes

- add `src/lib/registry.ts`
- replace `bin/lib/registry.js` with a thin shim to `dist/lib/registry`
- route registry persistence through the new `config-io` substrate
- update the write-failure test to accept the new user-facing config
write error

## Testing

- `npm run build:cli`
- `npx vitest run test/registry.test.js`
- `npx vitest run test/cli.test.js test/registry.test.js`

## Stack

Depends on #1552 (`refactor(cli): add config-io module`). Once that
merges, this PR can be retargeted to `main`.

Relates to #924.
